requirements.txt
blinker==1.6.2
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
